[BUGFIX] Use offset bin positions consistently
Handling of offset bin positions was not implemented consistently. This has
resulted in offset bin positions being used in plotting output, but not in
summary text output.

Change function `load_table` to return an array of bin edges, which is a numpy
array of integer bin start positions, followed by the end position of the last
bin (all in base pairs). Wherever bin starts only are used, these are explicitly
set from the `bins` array to a variable `bin_starts`.

Change function `doComputation` to accept bin edge positions, and to index into
these positions with the interval edge variables `left` and `right`. Bin edges
are needed because the `right` index may be `T` in some cases where an interval
extends beyond the rightmost bin. In such cases, the rightmost bin edge position
is used. Ensure left indices of intervals are greater than or equal to zero, and
right indices are less than or equal to `T`.

Change function `doPlotting` to set the horizontal plotting positions in `X`
from the mid-points of the offset bins, to draw the 90% Bayesian credible
interval between offset bin positions, and to plot the axis between the first
and last bin edges.